chore(ui): Move widget and empty widget for reuse
