+++ ui/apps/platform/src/Components/PatternFly/WidgetErrorEmptyState.tsx
